refactor(transport): move releases_rbac to transport/auth/dispatch_rbac.rs
Keeps transport/mod.rs under the 500-line structure cap; the dispatch-level
RBAC test lives beside the auth middleware it exercises.
